requirements/common.txt
# Core Stuff
# -------------------------------------
Django==1.10.5
django-flat-responsive==1.2.0

# WSGI HTTP Server
# -------------------------------------
uWSGI==2.0.14

# Configuration
# -------------------------------------
argon2-cffi==16.3.0
django-environ==0.4.1
django-sites==0.9
python-dotenv==0.6.2

# Staticfiles
# -------------------------------------
whitenoise==3.2.3

# Extensions
# -------------------------------------
pytz==2016.10

# Models
# -------------------------------------
psycopg2==2.6.2

Pillow==4.0.0
django-versatileimagefield==1.6.2
django-uuid-upload-path==1.0.0

# REST APIs
# -------------------------------------
djangorestframework==3.5.3

# LOGGING
# -------------------------------------
django-log-request-id==1.3.2

# Raven is the Sentry client
# --------------------------
raven==5.32.0
